BL-635 Cleaned up RenderTableContainer, moved data fetching to its parent container, linked its data content to the unified data store, fixed errors regarding missing keys for mapped components.
